Default values populate newgame.
