+++ client/src/components/button.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 const Button = (props) => {
   const handleClick = () => {
